backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
python-dotenv==1.0.0

# File Handling
boto3==1.34.34
Pillow==10.2.0
aiofiles==24.1.0

# Communication
twilio==8.11.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1